\name{GeometricMean}
\alias{GeometricMean}
\title{GeometricMean
}
\description{Calculates the geometric mean (the nth root of the product of n values).
}
\usage{
GeometricMean(x)
}
\arguments{
  \item{x}{Vector of n values whose geometric mean is to be calculated.
}
}
\value{The geometric mean of x.
}
\references{Ruscio et al. (2012)
}
\author{John Ruscio
}
\examples{
GeometricMean(c(1, 3, 9, 27, 81))
}


