\name{HarmonicMean}
\alias{HarmonicMean}
\title{HarmonicMean
}
\description{Calculates the harmonic mean (the reciprocal of the arithmetic mean of the reciprocals of n values).
}
\usage{
HarmonicMean(x)
}
\arguments{
  \item{x}{Vector of n values whose harmonic mean is to be calculated.
}
}
\value{The harmonic mean of x.
}
\references{
Ruscio et al. (2012)
}
\author{John Ruscio
}
\examples{HarmonicMean(c(1, 3, 9, 27, 81))
}
